requirements.txt
asgiref==3.8.1
atomicwrites==1.4.1
attrs==23.2.0
colorama==0.4.6
coverage==7.6.0
Django==5.0.7
django-bootstrap5==22.2
django-filter==24.2
django_debug_toolbar==3.8.1
djangorestframework==3.15.2
djangorestframework-simplejwt==4.8.0
flake8==5.0.4
flake8-docstrings==1.7.0
iniconfig==2.0.0
mccabe==0.7.0
mypkg==1.0.0
mypy-extensions==1.0.0
packaging==24.1
pep8-naming==0.13.3
pluggy==1.5.0
py==1.11.0
pyannotate==1.2.0
pycodestyle==2.9.1
pydocstyle==6.3.0
pyflakes==2.5.0
PyJWT==2.9.0
pytest==7.1.2
pytest-cov==5.0.0
pytest-django==4.5.2
pytest-lazy-fixture==0.6.3
pytz==2024.1
six==1.16.0
snowballstemmer==2.2.0
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.12.2
tzdata==2024.1
